requirements.txt
Flask
python-dotenv
PyMySQL
flask_login
flask_wtf